chore: ensure footer always loads
