+++ static/scripts/rewards/init.ts
@@ -6,14 +6,6 @@ grid(document.getElementById("grid") as HTMLElement, gridLoadedCallback); // @DE
 
 readClaimDataFromUrl(app).catch(console.error); // @DEV: read claim data from URL
 
-declare const commitHash: string; // @DEV: passed in at build time check build/esbuild-build.ts
-export function displayCommitHash() {
-  // display commit hash in footer
-  const buildElement = document.querySelector(`#build a`) as HTMLAnchorElement;
-  buildElement.innerHTML = commitHash;
-  buildElement.href = `https://github.com/ubiquity/pay.ubq.fi/commit/${commitHash}`;
-}
-
 // cSpell:ignore llback
 function gridLoadedCallback() {
   document.body.classList.add("grid-loaded");
