+++ static/scripts/rewards/render-transaction/display-commit-hash.ts
@@ -1,0 +1,7 @@
+declare const commitHash: string; // @DEV: passed in at build time check build/esbuild-build.ts
+export function displayCommitHash() {
+  // display commit hash in footer
+  const buildElement = document.querySelector(`#build a`) as HTMLAnchorElement;
+  buildElement.innerHTML = commitHash;
+  buildElement.href = `https://github.com/ubiquity/pay.ubq.fi/commit/${commitHash}`;
+}
